feat: frontend side login (#15)
Adds a login route to the application. Currently only frontend side
implementation is available and no backend integration was done.
Any username is allowed so long as the password is `test`. Once the user
is logged in no relogin is necessary and logging out is not possible as
it won't be useful for the players. For testing and development purposes
you can log out by removing the `currentUser` entry in localStorage

---------

Co-authored-by: Borys Juskiw